requirements.txt
numpy
flask
flask_cors
openai
discord
pytz
qrcode
requests
cloudinary
tiktoken
unidecode
elevenlabs
hashlib
python-dotenv
importlib